Log: changed init; TriggerTest: added mocks
Log needed Executive and Executive (used in TriggerTest) needed Log.
Because of this, they _both_ couldn't take the other in the constructor.
I moved the Executive assignment in Log out of the constructor.

I added Log and parent mocks in the TriggerTest.  The only one that's
there is the MessageTrigger but it's a start.

Adds logging through the Executive.